support moar heap snapshot format v2
it uses integers printed raw to a binary file, all records
except the string heap are uniformly sized, the file ends in a
table of segment sizes.

read the whole file with multiple file handles at the same time
in multiple threads for extra speed.

it is almost 2x faster and uses 1/3rd of the ram of format v1
for parsing, but the files are noticably bigger.