experimental impl of new heap snapshot format v3
there's a patch for moarvm that prints every
field of all the structs into separate files as
ascii numbers gzipped.

this script reads those files and spits out the
new file format.

Soon it'll also read the format and spit the same
files back out so they can be compared.